% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_spectrum_cache}
\alias{clear_spectrum_cache}
\title{Clear the spectrum cache}
\usage{
clear_spectrum_cache()
}
\value{
NULL (invisible)
}
\description{
Removes all cached spectra from memory. Useful when memory is low
or when spectra files have been modified on disk.
}
