% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalesce operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value to check}

\item{b}{Default value if a is NULL}
}
\value{
a if not NULL, otherwise b
}
\description{
Returns the first argument if it is not NULL, otherwise returns the second.
Similar to the ?? operator in C# or the // operator in Perl.
}
\examples{
NULL \%||\% "default"  # returns "default"
"value" \%||\% "default"  # returns "value"
}
