% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box_editor.R
\name{mod_box_editor_server}
\alias{mod_box_editor_server}
\title{Box Editor Module - Server}
\usage{
mod_box_editor_server(id, rv, data_reactives, parent_input, parent_session)
}
\arguments{
\item{id}{Character. The module's namespace ID}

\item{rv}{List. Shared reactive values}

\item{data_reactives}{List. Reactive expressions}

\item{parent_input}{Shiny input. Parent input object for table selections}

\item{parent_session}{Shiny session. Parent session for plotlyProxy}
}
\value{
NULL (side effects only)
}
\description{
Box Editor Module - Server
}
