% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_box_editor.R
\name{mod_box_editor_ui}
\alias{mod_box_editor_ui}
\title{Box Editor Module - UI}
\usage{
mod_box_editor_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing box editing controls
}
\description{
Box Editor Module - UI
}
