% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_click_mode.R
\name{mod_click_mode_ui}
\alias{mod_click_mode_ui}
\title{Click Mode Module - UI}
\usage{
mod_click_mode_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing click mode controls
}
\description{
Click Mode Module - UI
}
