% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_delete.R
\name{mod_delete_server}
\alias{mod_delete_server}
\title{Delete Module - Server}
\usage{
mod_delete_server(id, rv)
}
\arguments{
\item{id}{Character. The module's namespace ID}

\item{rv}{List. Shared reactive values (centroids_data, modifiable_boxes, fixed_boxes, pending_deletions)}
}
\value{
NULL (side effects only)
}
\description{
Delete Module - Server
}
