% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_delete.R
\name{mod_delete_ui}
\alias{mod_delete_ui}
\title{Delete Module - UI}
\usage{
mod_delete_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing delete controls
}
\description{
Delete Module - UI
}
