% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export.R
\name{mod_export_ui}
\alias{mod_export_ui}
\title{Export Module - UI}
\usage{
mod_export_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing export controls
}
\description{
Export Module - UI
}
