% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fusion.R
\name{mod_fusion_server}
\alias{mod_fusion_server}
\title{Fusion Module - Server}
\usage{
mod_fusion_server(id, rv)
}
\arguments{
\item{id}{Character. The module's namespace ID}

\item{rv}{List. Shared reactive values (centroids_data, modifiable_boxes, fixed_boxes, pending_fusions)}
}
\value{
NULL (side effects only)
}
\description{
Fusion Module - Server
}
