% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fusion.R
\name{mod_fusion_ui}
\alias{mod_fusion_ui}
\title{Fusion Module - UI}
\usage{
mod_fusion_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing fusion controls
}
\description{
Fusion Module - UI
}
