% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import.R
\name{mod_import_server}
\alias{mod_import_server}
\title{Import Module - Server}
\usage{
mod_import_server(id, rv, refresh_nmr_plot)
}
\arguments{
\item{id}{Character. The module's namespace ID}

\item{rv}{List. Shared reactive values}

\item{refresh_nmr_plot}{Function. Function to refresh the NMR plot}
}
\value{
NULL (side effects only)
}
\description{
Import Module - Server
}
