% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import.R
\name{mod_import_ui}
\alias{mod_import_ui}
\title{Import Module - UI}
\usage{
mod_import_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing import controls
}
\description{
Import Module - UI
}
