% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manual_add.R
\name{mod_manual_add_ui}
\alias{mod_manual_add_ui}
\title{Manual Add Module - UI}
\usage{
mod_manual_add_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing manual add controls
}
\description{
Manual Add Module - UI
}
