% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_manual_editing.R
\name{mod_manual_editing_ui}
\alias{mod_manual_editing_ui}
\title{Manual Editing Module - UI (Wrapper)}
\usage{
mod_manual_editing_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing all manual editing UI components
}
\description{
Manual Editing Module - UI (Wrapper)
}
