% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pending_changes.R
\name{mod_pending_changes_ui}
\alias{mod_pending_changes_ui}
\title{Pending Changes Module - UI}
\usage{
mod_pending_changes_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing apply/discard buttons
}
\description{
Pending Changes Module - UI
}
