% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reset.R
\name{mod_reset_server}
\alias{mod_reset_server}
\title{Reset Module - Server}
\usage{
mod_reset_server(id, status_msg, rv, parent_session)
}
\arguments{
\item{id}{Character. The module's namespace ID}

\item{status_msg}{ReactiveVal. Shared status message reactive value}

\item{rv}{List. Shared reactive values}

\item{parent_session}{Shiny session. Parent session for updating inputs}
}
\value{
A list containing:
\itemize{
\item \code{reset_triggered}: ReactiveVal that increments when reset is triggered
}
}
\description{
Reset Module - Server
}
