% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reset.R
\name{mod_reset_ui}
\alias{mod_reset_ui}
\title{Reset Module - UI}
\usage{
mod_reset_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing reset button
}
\description{
Reset Module - UI
}
