% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_export.R
\name{mod_save_export_server}
\alias{mod_save_export_server}
\title{Save & Export Module - Server (Wrapper)}
\usage{
mod_save_export_server(
  id,
  status_msg,
  rv,
  load_data,
  data_reactives,
  refresh_nmr_plot,
  parent_session,
  parent_input
)
}
\arguments{
\item{id}{Character. The module's namespace ID}

\item{status_msg}{ReactiveVal. Shared status message reactive value}

\item{rv}{List. Named list of reactive values}

\item{load_data}{List. Return value from mod_load_data_server}

\item{data_reactives}{List. Named list of reactive expressions}

\item{refresh_nmr_plot}{Function. Function to refresh the NMR plot}

\item{parent_session}{Shiny session. Parent session for updating inputs}

\item{parent_input}{Shiny input. Parent input object for reading UI values}
}
\value{
A list containing:
\itemize{
\item \code{reset_triggered}: Reactive that increments when reset is triggered
}
}
\description{
Save & Export Module - Server (Wrapper)
}
