% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_save_export.R
\name{mod_save_export_ui}
\alias{mod_save_export_ui}
\title{Save & Export Module - UI (Wrapper)}
\usage{
mod_save_export_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing all save/export UI components
}
\description{
Save & Export Module - UI (Wrapper)
}
