% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_session.R
\name{mod_session_server}
\alias{mod_session_server}
\title{Session Module - Server}
\usage{
mod_session_server(
  id,
  status_msg,
  rv,
  load_data,
  refresh_nmr_plot,
  parent_session,
  parent_input
)
}
\arguments{
\item{id}{Character. The module's namespace ID}

\item{status_msg}{ReactiveVal. Shared status message reactive value}

\item{rv}{List. Shared reactive values}

\item{load_data}{List. Return value from mod_load_data_server}

\item{refresh_nmr_plot}{Function. Function to refresh the NMR plot}

\item{parent_session}{Shiny session. Parent session for updating inputs}

\item{parent_input}{Shiny input. Parent input object for reading UI values}
}
\value{
NULL (side effects only)
}
\description{
Session Module - Server
}
