% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_session.R
\name{mod_session_ui}
\alias{mod_session_ui}
\title{Session Module - UI}
\usage{
mod_session_ui(id)
}
\arguments{
\item{id}{Character. The module's namespace ID}
}
\value{
A tagList containing session save/load controls
}
\description{
Session Module - UI
}
